- Add support for managing non-prolog exceptions
When gprolog is used within a process that has may itself provoke exceptions, the global handler installed here would exit() the application.
The function set_chain_nonhandled_exceptions when set to TRUE allows for those exceptions to be handled by the previous exception handler.
For example, if some thread was executing some unrelated code that generated an exception (e.g. dereferencing a NULL pointer), the prolog exception handler has the first chance to manage the exception and if it remains unhandled it will be left to be handled by the previous handler instead of exiting the process.